Refactor #Etienne
Everything is an ExecutableStep.
A Sequence, a GotoStep and a Delivery directly herit from ExecutableStep
A Tour does not exist anymore It is a Sequence of executableStep
(sequence).
A DropPoint does not exist anymore. It is a Sequence of executableStep
(a Goto and a sequence)
=> The key is that is it the assembly that will build the correct
sequence. That is all. <3
Delete TourState, Tour, Step and dropPoint